GSANet/requirements.txt
accimage==0.2.0
ADEval==1.1.0
apex==0.9.10dev
einops==0.8.2
fvcore==0.1.5.post20221221
imgaug==0.4.0
matplotlib==3.7.5
numba==0.58.1
numpy==1.24.1
opencv_python==4.10.0.84
Pillow==8.4.0
Pillow==12.1.0
scikit_learn==1.3.2
scipy==1.17.0
skimage==0.0
tabulate==0.9.0
tensorboardX==2.6.2.2
tensorboardX==2.6.4
timm==0.8.15.dev0
torch==2.1.2+cu118
torchvision==0.16.2+cu118
